check values now

+++ assets/js/custom.js
@@ -448,8 +448,8 @@ $(document).ready(function () {
          staffId = $('#staffSelect').val();
         var amntPaid = $("#amntPaid").text;
         var creditAmnt = $("#creditToPay").text;
-        amntPaid = parseFloat(amntPaid);
-        creditAmnt = parseFloat(creditAmnt);
+        amntPaid = parseInt(amntPaid);
+        creditAmnt = parseInt(creditAmnt);
         //console.log(randomCustomerName+ "name");
         //console.log(billDiscount+''+billTotal+''+billAmountPayable+''+customerType+''+customerId+''+randomCustomerName+''+staffId);
          if((randomCustomerName != '' || customerId != '') && staffId != ''  && billTotal != '0'){
